Reorganized test folder
